requirements.txt
transformers>=4.44.0
accelerate>=0.34.0
datasets>=2.20.0
peft>=0.12.0
torch>=2.2.0,<2.5.0
torchvision>=0.17.0,<0.20.0
sentencepiece>=0.2.0
scikit-learn>=1.4.0
faiss-cpu>=1.7.4
numpy>=1.24
pydantic>=2.6
protobuf>=4.25.0
rich>=13.7
typer>=0.12
pytest>=7.4


